# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClawOPT backend",
    "version": "1.0.0",
    "description": "Generated from the backend route registry (npm run openapi:generate). Do not edit by hand."
  },
...
  "paths": {
...
    "/api/agents/{agentId}/workspace-files": {
      "get": {
        "operationId": "getApiAgentsAgentIdWorkspaceFiles",
        "tags": [
          "control/workspace-files"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/agents/:agentId/workspace-files",
        "x-clawopt-order": 129
      }
    },
    "/api/agents/{agentId}/workspace-files/{name}": {
      "get": {
        "operationId": "getApiAgentsAgentIdWorkspaceFilesName",
        "tags": [
          "control/workspace-files"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/agents/:agentId/workspace-files/:name",
        "x-clawopt-order": 130
      },
      "put": {
        "operationId": "putApiAgentsAgentIdWorkspaceFilesName",
        "tags": [
          "control/workspace-files"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/agents/:agentId/workspace-files/:name",
        "x-clawopt-order": 131
      }
    },
    "/api/write-gate/settings": {
      "get": {
        "operationId": "getApiWriteGateSettings",
        "tags": [
          "control/write-gate"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/write-gate/settings",
        "x-clawopt-order": 132
      }
    },
    "/api/write-gate/settings/{agentId}": {
      "put": {
        "operationId": "putApiWriteGateSettingsAgentId",
        "tags": [
          "control/write-gate"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/write-gate/settings/:agentId",
        "x-clawopt-order": 133
      }
    },
    "/api/write-gate/pending": {
      "get": {
        "operationId": "getApiWriteGatePending",
        "tags": [
          "control/write-gate"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/write-gate/pending",
        "x-clawopt-order": 134
      }
    },
    "/api/write-gate/pending/{id}": {
      "get": {
        "operationId": "getApiWriteGatePendingId",
        "tags": [
          "control/write-gate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/write-gate/pending/:id",
        "x-clawopt-order": 135
      }
    },
    "/api/write-gate/pending/{id}/approve": {
      "post": {
        "operationId": "postApiWriteGatePendingIdApprove",
        "tags": [
          "control/write-gate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/write-gate/pending/:id/approve",
        "x-clawopt-order": 136
      }
    },
    "/api/write-gate/pending/{id}/reject": {
      "post": {
        "operationId": "postApiWriteGatePendingIdReject",
        "tags": [
          "control/write-gate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/write-gate/pending/:id/reject",
        "x-clawopt-order": 137
      }
    },
...
    "/api/plugins": {
      "get": {
        "operationId": "getApiPlugins",
        "tags": [
          "control/plugins"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/plugins",
        "x-clawopt-order": 170
      }
    },
    "/api/plugins/{id}": {
      "get": {
        "operationId": "getApiPluginsId",
        "tags": [
          "control/plugins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/plugins/:id",
        "x-clawopt-order": 171
      },
      "delete": {
        "operationId": "deleteApiPluginsId",
        "tags": [
          "control/plugins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/plugins/:id",
        "x-clawopt-order": 176
      }
    },
    "/api/plugins/{id}/enable": {
      "post": {
        "operationId": "postApiPluginsIdEnable",
        "tags": [
          "control/plugins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/plugins/:id/enable",
        "x-clawopt-order": 172
      }
    },
    "/api/plugins/{id}/disable": {
      "post": {
        "operationId": "postApiPluginsIdDisable",
        "tags": [
          "control/plugins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/plugins/:id/disable",
        "x-clawopt-order": 173
      }
    },
    "/api/plugins/install": {
      "post": {
        "operationId": "postApiPluginsInstall",
        "tags": [
          "control/plugins"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/plugins/install",
        "x-clawopt-order": 174
      }
    },
    "/api/plugins/{id}/update": {
      "post": {
        "operationId": "postApiPluginsIdUpdate",
        "tags": [
          "control/plugins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/plugins/:id/update",
        "x-clawopt-order": 175
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "clawopt_session",
        "description": "httpOnly session cookie issued by POST /api/auth/login (web UI)."
      },
      "authTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ClawOPT-Auth-Token",
        "description": "Session token for CLI and scripts."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same session token sent as Authorization: Bearer."
      }
    }
  }
}